+++ src-tauri/hg-gui/src/lib.rs
@@ -157,6 +157,7 @@ pub fn run() {
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_process::init())
+        .plugin(tauri_plugin_notification::init())
         .plugin(tauri_plugin_updater::Builder::new().build())
         .invoke_handler(serve::wrap_invoke_handler(specta_builder.invoke_handler()))
         .build(tauri::generate_context!())